[Common Frontend] Move ComplexTypeMark into common frontend
